+++ app/serializers/static_invite_serializer.rb
@@ -1,0 +1,4 @@
+class StaticInviteSerializer < ActiveModel::Serializer
+  attributes :id, :static_id, :status
+  belongs_to :character, serializer: CharacterCrafterSerializer
+end