+++ frontend/components_react/static_management/static_management.jsx
@@ -18,6 +18,8 @@ export default class StaticManagement extends React.Component {
     this.state = {
       members: [],
       memberIds: [],
+      invites: [],
+      inviteIds: [],
       query: '',
       searchedCharacters: []
     }
@@ -40,7 +42,10 @@ export default class StaticManagement extends React.Component {
         const memberIds = data.characters.map((character) => {
           return character.id
         })
-        this.setState({members: data.characters, memberIds: memberIds})
+        const inviteIds = data.invites.map((invite) => {
+          return invite.character.id
+        })
+        this.setState({members: data.characters, memberIds: memberIds, invites: data.invites, inviteIds: inviteIds})
       }
     })
   }
@@ -50,11 +55,36 @@ export default class StaticManagement extends React.Component {
       method: 'GET',
       url: `/api/v1/characters/search.json?access_token=${this.props.access_token}&query=${this.state.query}&world_id=${this.props.world_id}&fraction_id=${this.props.fraction_id}`,
       success: (data) => {
-        console.log(data)
         const characters = data.characters.filter((character) => {
-          return !this.state.memberIds.includes(character.id)
+          return !this.state.memberIds.includes(character.id) && !this.state.inviteIds.includes(character.id)
         })
         this.setState({searchedCharacters: characters})
+      }
+    })
+  }
+
+  _onInviteCharacter(character) {
+    $.ajax({
+      method: 'POST',
+      url: `/api/v1/static_invites.json?access_token=${this.props.access_token}`,
+      data: { static_invite: { static_id: this.props.static_id, character_id: character.id } },
+      success: (data) => {
+        const searchedCharacters = [... this.state.searchedCharacters]
+        const searchedCharacterIndex = searchedCharacters.indexOf(character)
+        searchedCharacters.splice(searchedCharacterIndex, 1)
+        if (data.character !== undefined) {
+          let members = this.state.members
+          members.push(data.character)
+          let memberIds = this.state.memberIds
+          memberIds.push(data.character.id)
+          this.setState({members: members, memberIds: memberIds, searchedCharacters: searchedCharacters})
+        } else if (data.invite !== undefined) {
+          let invites = this.state.invites
+          invites.push(data.invite)
+          let inviteIds = this.state.inviteIds
+          inviteIds.push(data.invite.character.id)
+          this.setState({invites: invites, inviteIds: inviteIds, searchedCharacters: searchedCharacters})
+        }
       }
     })
   }
@@ -73,6 +103,43 @@ export default class StaticManagement extends React.Component {
     })
   }
 
+  _renderInvites() {
+    if (this.state.invites.length === 0) return <p>No invites</p>
+    else {
+      return (
+        <table className="table table-sm">
+          <thead>
+            <tr>
+              <th>{strings.name}</th>
+              <th>{strings.race}</th>
+              <th>{strings.level}</th>
+              <th>{strings.guild}</th>
+              <th></th>
+            </tr>
+          </thead>
+          <tbody>
+            {this._renderInvitesResults()}
+          </tbody>
+        </table>
+      )
+    }
+  }
+
+  _renderInvitesResults() {
+    return this.state.invites.map((invite) => {
+      return (
+        <tr className={invite.character.character_class.en} key={invite.id}>
+          <td>{invite.character.name}</td>
+          <td>{invite.character.race[this.props.locale]}</td>
+          <td>{invite.character.level}</td>
+          <td>{invite.character.guild}</td>
+          <td>
+          </td>
+        </tr>
+      )
+    })
+  }
+
   _renderSearchedCharacters() {
     if (this.state.searchedCharacters.length > 0) {
       return (
@@ -83,6 +150,7 @@ export default class StaticManagement extends React.Component {
               <th>{strings.race}</th>
               <th>{strings.level}</th>
               <th>{strings.guild}</th>
+              <th></th>
             </tr>
           </thead>
           <tbody>
@@ -101,6 +169,9 @@ export default class StaticManagement extends React.Component {
           <td>{character.race[this.props.locale]}</td>
           <td>{character.level}</td>
           <td>{character.guild}</td>
+          <td>
+            <input type="submit" name="commit" value='Invite' className="btn btn-primary btn-sm" onClick={this._onInviteCharacter.bind(this, character)} />
+          </td>
         </tr>
       )
     })
@@ -141,6 +212,7 @@ export default class StaticManagement extends React.Component {
         <div className="double_line">
           <div className="form-group invites">
             <h3>{strings.invites}</h3>
+            {this._renderInvites()}
           </div>
           <div className="form-group search">
             <h3>{strings.search}</h3>