+++ spec/api/v1/statics_controller_spec.rb
@@ -270,4 +270,63 @@ RSpec.describe 'Statics API' do
       get '/api/v1/statics/form_values.json', headers: headers
     end
   end
+
+  describe 'GET#members' do
+    it_behaves_like 'API auth without token'
+    it_behaves_like 'API auth with invalid token'
+
+    context 'for logged user' do
+      let!(:user) { create :user }
+      let!(:guild) { create :guild }
+      let!(:character) { create :character, guild: guild, user: user }
+      let(:access_token) { JwtService.new.json_response(user: user)[:access_token] }
+
+      context 'for unexisted static' do
+        before { get '/api/v1/statics/unexisted/members.json', params: { access_token: access_token } }
+
+        it 'returns status 400' do
+          expect(response.status).to eq 400
+        end
+
+        it 'and returns error message' do
+          expect(JSON.parse(response.body)).to eq('error' => 'Object is not found')
+        end
+      end
+
+      context 'for existed static' do
+        let!(:static) { create :static, staticable: guild }
+
+        context 'for unavailable static' do
+          before { get "/api/v1/statics/#{static.id}/members.json", params: { access_token: access_token } }
+
+          it 'returns status 400' do
+            expect(response.status).to eq 400
+          end
+
+          it 'and returns error message' do
+            expect(JSON.parse(response.body)).to eq('error' => 'Forbidden')
+          end
+        end
+
+        context 'for available static' do
+          let!(:guild_role) { create :guild_role, guild: guild, character: character, name: 'gm' }
+          before { get "/api/v1/statics/#{static.id}/members.json", params: { access_token: access_token } }
+
+          it 'returns status 200' do
+            expect(response.status).to eq 200
+          end
+
+          %w[characters invites].each do |attr|
+            it "and #{attr}" do
+              expect(response.body).to have_json_path(attr)
+            end
+          end
+        end
+      end
+    end
+
+    def do_request(headers = {})
+      get '/api/v1/statics/unexisted/members.json', params: { static: { name: '1' } }, headers: headers
+    end
+  end
 end