+++ spec/interactors/create_static_member_spec.rb
@@ -3,14 +3,29 @@ describe CreateStaticMember do
   let!(:static) { create :static, world: character.world, fraction: character.race.fraction, staticable: character }
 
   describe '.call' do
-    let(:interactor) { CreateStaticMember.call(static: static, character: character) }
+    context 'for existed member' do
+      let!(:static_member) { create :static_member, static: static, character: character }
+      let(:interactor) { CreateStaticMember.call(static: static, character: character) }
 
-    it 'succeeds' do
-      expect(interactor).to be_a_success
+      it 'fails' do
+        expect(interactor).to be_a_failure
+      end
+
+      it 'and does not create static member' do
+        expect { interactor }.to_not change(static.static_members, :count)
+      end
     end
 
-    it 'and creates static member' do
-      expect { interactor }.to change { static.static_members.count }.by(1)
+    context 'for unexisted member' do
+      let(:interactor) { CreateStaticMember.call(static: static, character: character) }
+
+      it 'succeeds' do
+        expect(interactor).to be_a_success
+      end
+
+      it 'and creates static member' do
+        expect { interactor }.to change { static.static_members.count }.by(1)
+      end
     end
   end
 end