+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - api endpoint characters#search
 - rendering search results
 - StaticInvite model
+- creating invites to statics
 
 ## [0.6.0] - 2019-11-01
 ### Added